// types/state.ts
import Dialog from "../components/ui/Dialog"
export enum EventType {
  Default = "default",
  Primary = "primary",
  Success = "success",
  Warning = "warning",
  Danger = "danger",
  Icon = "icon",
}

export enum ButtonAction {
  Navigate = "navigate",
  Submit = "submit",
}

export enum DialogRole {
  Dialog = "dialog",
  AlertDialog = "alertdialog",
}

export enum Size {
  Small = "small",
  Medium = "medium",
  Large = "large",
}